requirements.txt
# Lehigh Mars Rover AI Agent - Separate Environment
fastapi
uvicorn
python-dotenv
pytest
pinecone
openai
PyMuPDF
